+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Use FactoryBot instead of FactoryGirl
+
 ## [1.0.0] - 2016-02-14
 ### Changed
 - Relax Rails constraint to `~> 4.2`